==> 01-personnel_checkins_data.ipynb <==
# Dunham's Data

## Setting up

# from datetime import datetime

# import networkx as nx
import pandas as pd
# import seaborn as sns

pd.options.display.max_rows = 1000
pd.options.display.max_columns = 1000

## Loading data

df = pd.read_excel('data/src/DunhamsData_PersonnelCheckIns_DatasetSample.xlsx')
df.fillna('', inplace=True)

df.columns

df.shape

df.head()

## Processing data

df['Comprehensive Check In'].unique()

df2 = df[df['Comprehensive Check In'] != '']
df2 = df2.set_index('Date')
df2.index = [x.strftime('%b %d, %Y') for x in df2.index]
df2.index.name = None

N = 5
# df2 = df2.iloc[:, N:]

df2.shape

df2.head()

yn_set = set()
for c in df2.iloc[:, N:]:
    yn_set = yn_set.union(df2[c].unique())

yn_set

df3 = df2.copy()
df3.iloc[:, N:] = df3.iloc[:, N:].replace('  ', '')

df3.head()

# df2.iloc[:, N:] = df2.iloc[:, N:].replace('n*', 'n')
df2.iloc[:, N:] = df2.iloc[:, N:].replace('y*', 'y')
df2.iloc[:, N:] = df2.iloc[:, N:].replace('  ', '')
df2.iloc[:, N:] = df2.iloc[:, N:].replace('?', 'n')

yn_set = set()
for c in df2.iloc[:, N:]:
    yn_set = yn_set.union(df2[c].unique())

yn_set

# Check there is no extra blank spaces in the names
checkin_names = df2.columns[N:]
for name in checkin_names:
    if ',' in name:
        first = name.split(',')[1].strip()
        last = name.split(',')[0].strip()
        fullname = last + ', ' + first
    else:
        fullname = name.strip()
    assert(fullname == name)

# %%time

for i, (j, row) in enumerate(df2.iterrows()):
    if row['Comprehensive Check In'] == 'y' or i == 0:
        # df2.loc[j].iloc[N:] = row[N:].replace('', 'n')
        df2.iloc[i, N:] = row[N:].replace('', 'n')
    else:
        prev = df2.iloc[i - 1]
        for k, x in enumerate(row):
            if x == '':
                df2.iloc[i, k] = prev[k]

df2.shape

df2

# Count values for confidence column
df2.confidence.value_counts()

## Saving data

df2.to_csv('data/out/checkins.csv')
