==> 03-city_and_country.ipynb <==
# from datetime import datetime

# # import community as community2 # pip install python-louvain
# import networkx as nx
# from networkx.algorithms import community
import pandas as pd
# import seaborn as sns

pd.options.display.max_rows = 1000
pd.options.display.max_columns = 1000

## Loading data

dunham_df = pd.read_csv('data/out/dunham_1947-60.csv')
dunham_df.fillna('', inplace=True)

dunham_df.shape

dunham_df.head()

## Processing data

def pick_city(row):
    if row['CITY1'] and 'transit' not in row['CITY1'] and '[[' not in row['CITY1']:
        r = row['CITY1']
    elif row['CITY2'] and 'transit' not in row['CITY2'] and '[[' not in row['CITY2']:
        r = row['CITY2']
    else:
        r = ''
    return r

def pick_country(row):
    if row['CITY1'] and 'transit' not in row['CITY1'] and '[[' not in row['CITY1']:
        r = row['COUNTRY1']
    elif row['CITY2'] and 'transit' not in row['CITY2'] and '[[' not in row['CITY2']:
        r = row['COUNTRY2']
    else:
        r = ''
    return r

dunham_df2 = dunham_df[['DATE', 'CITY1', 'COUNTRY1', 'CITY2', 'COUNTRY2']].copy()
dunham_df2['CITY'] = dunham_df2.apply(pick_city, axis=1)
dunham_df2['COUNTRY'] = dunham_df2.apply(pick_country, axis=1)
dunham_df2 = dunham_df2[dunham_df2.CITY != ''][['DATE', 'CITY', 'COUNTRY']]
dunham_df2.set_index('DATE', inplace=True)
dunham_df2.index.name = None

dunham_df2.shape

dunham_df2.head()

## Saving data

dunham_df2.to_csv('data/out/everyday_city_country.csv')
