==> 04-public_akas.ipynb <==
# Dunham's Data

## Setting up

import numpy as np
import pandas as pd
import xlrd

pd.options.display.max_rows = 1000
pd.options.display.max_columns = 1000

## Loading data

### Checkins

checkins_df = pd.read_csv('data/out/checkins.csv', index_col=0)

checkins_df.columns

N = 5
len(checkins_df.columns[N:])

### Personnel

personnel_df = pd.read_excel('data/src/DunhamsData_PersonnelAttributes_DatasetSample.xlsx')
personnel_df.fillna('', inplace=True)

personnel_df.shape

personnel_df.head()

personnel_df.tail()

## Pre-processing data

set(checkins_df.columns[N:]) == set(personnel_df.Name.unique())

def format_fullname(name):
    assert(',' not in name or ', ' in name)
    if ', ' in name:
        last, first = name.split(', ')
        r = first + ' ' + last
    else:
        r = name
    return r

## Processing data

def str2set(s):
    return set(x.strip() for x in s.split(';') if x.split())

def set2str(s):
    return ' / '.join(s)

public_akas_df = pd.DataFrame(columns=['SCREEN NAME', 'PUBLIC AKAS'])
for i, row in personnel_df.iterrows():
    name = row['Name']
    akas = row['public akas']
    screen_name = format_fullname(name)
    public_akas = str2set(akas)
    public_akas = screen_name + ' / ' + set2str(public_akas) if public_akas else screen_name
    public_akas_df.loc[i] = [screen_name, public_akas]

public_akas_df

## Saving data

public_akas_df.to_csv('web/data/public_akas.csv', index=False)
